+++ integration_tests/test_os.py
@@ -2,13 +2,13 @@ from ltypes import i64
 from os import (open, read, close, O_RDONLY)
 
 def test():
-    path: str 
+    path: str
     path = "../test_os.py"
     fd: i64
     n: i64
     fd = open(path, O_RDONLY)
-    n = 100
+    n = i64(100)
     print(read(fd, n))
-    close(fd) 
+    close(fd)
 
 test()
